requirements.txt
flask
xgboost
catboost
scikit-learn
imbalanced-learn
numpy
pandas
joblib
gunicorn
pymysql
sqlalchemy
flask-sqlalchemy
python-dotenv
alembic